Move restic source to mover file
